try-catch blocks added

+++ CASWebApi/Controllers/StudentController.cs
@@ -25,22 +25,35 @@ namespace CASWebApi.Controllers
         /// </summary>
         /// <returns>List of Student</returns>
         [HttpGet("getAllStudents", Name = nameof(GetAllStudents))]
-        public ActionResult<List<Student>> GetAllStudents() =>
-             _studentService.GetAll();
+        public ActionResult<List<Student>> GetAllStudents()
+        {
+            try
+            {
+                return _studentService.GetAll();
+            }
+            catch(Exception e)
+            {
+                return BadRequest("No connection to database");
+            }
+        }
 
+        /// <summary>
+        /// get student list by groupNumber
+        /// </summary>
+        /// <param name="groupName"></param>
+        /// <returns>list of students,or error message if not found or was exception</returns>
         [HttpGet("getAllStudentsByGroup", Name = nameof(GetAllStudentsByGroup))]
-
         public ActionResult<List<Student>> GetAllStudentsByGroup(string groupName)
         {
             try
             {
-                var students = _studentService.GetAllStudentsByGroup(groupName);
-
-                if (students == null)
+                if (groupName != null)
                 {
-                    return NotFound();
+                    var students = _studentService.GetAllStudentsByGroup(groupName);
+                        return students;
                 }
-                return students;
+                return BadRequest("groupName was null");
+
             }
             catch (Exception e)
             {
@@ -49,6 +62,11 @@ namespace CASWebApi.Controllers
 
         }
 
+        /// <summary>
+        /// get all students by faculties names
+        /// </summary>
+        /// <param name="facultyNames"></param>
+        /// <returns>list of students,or error message if not found or was exception</returns>
         [HttpGet("GetAllStudentsByFaculties", Name = nameof(GetAllStudentsByFaculties))]
         public ActionResult<List<Student>> GetAllStudentsByFaculties([FromQuery] string[] facultyNames)
         {
@@ -64,7 +82,7 @@ namespace CASWebApi.Controllers
                     return BadRequest("No connection to database");
                 }
             }
-            return NotFound("facultyNames param is null ");
+            return BadRequest("facultyNames param is null ");
 
 
 
@@ -84,7 +102,6 @@ namespace CASWebApi.Controllers
                 try
                 {
                     var students = _studentService.GetAllStudentsByGroups(groupNames);
-                    if (students != null)
                         return students;
                 }
                 catch (Exception e)
@@ -92,7 +109,7 @@ namespace CASWebApi.Controllers
                     return BadRequest("No connection to database");
                 }
             }
-            return NotFound("groupNames param is null");
+            return BadRequest("groupNames param is null");
 
         }
 
@@ -113,15 +130,29 @@ namespace CASWebApi.Controllers
                 return BadRequest("No connection to database");
             }
         }
-       
+
         /// <summary>
         /// Get number of Students by class
         /// </summary>
         /// <param name="id">Id of group</param>
         /// <returns>Number of student in specific group</returns>
         [HttpGet("getNumberOfStudentsInClass", Name = nameof(GetNumberOfStudentsInClass))]
-        public ActionResult<int> GetNumberOfStudentsInClass(string id) =>
-             _studentService.GetNumberOfStudentsByClass(id);
+        public ActionResult<int> GetNumberOfStudentsInClass(string id)
+        {
+            if (id != null)
+            {
+                try
+                {
+                    return _studentService.GetNumberOfStudentsByClass(id);
+                }
+                catch
+                {
+                    return BadRequest("No connection to database");
+                }
+            }
+            else
+                return BadRequest("given id is null");
+        }
 
         /// <summary>
         /// Get Student profile by Id
@@ -131,21 +162,24 @@ namespace CASWebApi.Controllers
         [HttpGet("getStudentById", Name = nameof(GetStudentById))]
         public ActionResult<Student> GetStudentById(string id)
         {
-            try 
+            if (id != null)
             {
-                var student = _studentService.GetById(id);
-
-                if (student == null)
-                    return NotFound();
-
-
-                return student;
+                try
+                {
+                    var student = _studentService.GetById(id);
+                    if (student == null)
+                        return NotFound("student with given id doesn't exists in db");
+                    return student;
+                }
+                catch (Exception e)
+                {
+                    return BadRequest("No connection to database");
+                }
             }
-            catch(Exception e)
-            {
-                return BadRequest("No connection to database");
-            }
-            }
+            else
+             return BadRequest("given id was null");
+
+        }
         
 
         /// <summary>
@@ -160,20 +194,16 @@ namespace CASWebApi.Controllers
             {
                 try
                 {
-                    if (_studentService.Create(student))
-                    {
-                        return CreatedAtRoute("getStudentById", new { id = student.Id }, student);
-                    }
-                    else
-                        return NotFound("duplicated id or wrong id format");
+                    _studentService.Create(student);               
+                       return CreatedAtRoute("getStudentById", new { id = student.Id }, student);       
                 }
                 catch (Exception e)
                 {            
-                        return Conflict(e);
+                    return Conflict(e);
                 }
             }
             else
-                return BadRequest(false);
+                return BadRequest("Student model is not correct");
             
         }
 
@@ -186,11 +216,21 @@ namespace CASWebApi.Controllers
         [HttpPost("insertListOfStudents", Name = nameof(InsertListOfStudents))]
         public ActionResult<Student> InsertListOfStudents(List<Student> students)
         {
-           
-            if (!(_studentService.InsertManyStudents(students)))
-                return NotFound("duplicated id or wrong id format");
-
-            return Ok(true);
+            if (students != null)
+            {
+                try
+                {
+                    if (!(_studentService.InsertManyStudents(students)))
+                        return NotFound(false);
+                    return Ok(true);
+                }
+                catch (Exception e)
+                {
+                    return Conflict(e);
+                }
+            }
+            else
+                return BadRequest("given list of students was null");
         }
 
         /// <summary>
@@ -201,19 +241,27 @@ namespace CASWebApi.Controllers
         [HttpPut("updateStudent", Name = nameof(UpdateStudent))]
         public IActionResult UpdateStudent(Student studentIn)
         {
-            
-            bool res=false;
-
-            var student = _studentService.GetById(studentIn.Id);
-
-            if (student == null)
+            if (studentIn.Id != null || studentIn != null)
             {
-                return NotFound();
-            }
-            res = _studentService.Update(studentIn.Id, studentIn);
-   
+                try
+                {
+                    bool res = false;
+                    var student = _studentService.GetById(studentIn.Id);
 
-            return Ok(res);
+                    if (student == null)
+                    {
+                        return NotFound();
+                    }
+                    res = _studentService.Update(studentIn.Id, studentIn);
+                    return Ok(res);
+                }
+                catch(Exception e)
+                {
+                    return BadRequest("No connection to database");
+                }
+            }
+            else
+                return BadRequest("Id of given student,or student object itself is null");
         }
 
         /// <summary>
@@ -224,12 +272,24 @@ namespace CASWebApi.Controllers
         [HttpDelete("deleteStudentById", Name = nameof(DeleteStudentById))]
         public IActionResult DeleteStudentById(string id)
         {
-            var student = _studentService.GetById(id);
-            if (student != null && _studentService.RemoveById(student.Id) )
-
-                return Ok(true);
-            
-            return NotFound();
+            if (id != null)
+            {
+                try
+                {
+                    var student = _studentService.GetById(id);
+                    if (student != null && _studentService.RemoveById(student.Id))
+                        return Ok(true);
+                    return NotFound("student object with given id doesn't exists");
+                }
+                catch (Exception e)
+                {
+                    return BadRequest("No Connection to database");
+                }
+            }
+            else
+            {
+                return BadRequest("given id was null");
+            }
         }
     }
 }
